pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pyinstaller>=6.15.0",
    "pyzbar>=0.1.9",
    "qrcode>=8.2",
    "streamlit>=1.48.1",
]
